Convert to single cluster mode
The manager can only manage a single Proxmox cluster, but can use
multiple nodes when interacting with the cluster. Specify all nodes in
the cluster, so that the manager can work even if some nodes are down or
unavailable.

It should be easy to run multiple instances on different ports for each cluster.